mac80211: Fix user allowed tpc value
Value 0 is now a valid tpc. This will revert the fixed tx power to auto tx
power for the station.

Currently, 44 dBm of tx power happens to be the maximum value listed in
regulatory domain database.  for 60 GHz. To be future proof, increase the range
of allowed max value of TPC from 30 to 70. Higher values are anyways by lower
layers depending on the regulatory domain.

Change-Id: Ie5086f47fcb0f897b950eedd314b938fc2bebf14
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>