mac80211: Avoid overriding beacon Tx for non-p2p ROC
If want to use remain-on-channel interface for off-channel scanning
with wmi 10.2.4/10.4 in AP mode, passing a huge dwell time as big as
10 secs (which was added to fix some issues in p2p) has a potential
to miss beacon Tx depending on where we are with respect to the next
TBTT by switching to the requested channel right away. Beacon miss
during off-channel may not be preferred especially when the requested
off-channel (ROC) duration is less than the beacon interval. Make the
work around of passing very high dwell time specific to p2p because
the work around is added for p2p issue. For non-p2p, Pass the requested
off-channel duration in dwell time. Also defer scan stop worker
by some more time to account for the time delay to complete the scan.

Change-Id: I537b2da9dc98e2a4353ddd68f862610cf1d1077b
Signed-off-by: Vasanthakumar Thiagarajan <vthiagar@codeaurora.org>