mac80211: Fix wifi script - error handling
Wifi script will return 1 if fst.sh script file doesn't exist. On the
other hand fst.sh is an optional script and it is normally not required,
therefore it shouldn't be a reason for returning of an error.

Fix by adding return values of 0 at the end of both methods where the file
would be included in case it existed, so that the last executed line would
no longer be an unsuccesfull check for the file existence.

Change-Id: If99bf0aaa15c15f3a5826ed1dd883e14da70c6f0
Signed-off-by: Marek Kwaczynski <mkwacz@codeaurora.org>