ath10k: fix reporting channel survey data
When user requests for survey dump data, driver is providing wrong survey
information. This information we sent is the survey data that we have
collected during previous user request.

This issue occurs because we request survey dump for wrong channel. With
this change, we correctly display the correct and current survey
information to userspace.

Change-Id: I76939097df988f563874554f27fa286b4b2b5fad
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>