wil6210: low level RF sector API
Added vendor commands for low level control over
RF sectors. It allows user space a fine-grained control
over RF characteristics for TX and RX, such as direction
and gain of TX/RX. Main usages are debugging and diagnostics,
but also operational use cases.
This API exposes the internals of the wil6210 RF, as such
it is not possible to use generic nl80211 API. Since
it is required for operational use cases debugfs is also
not suitable, so we opted for using vendor commands.
API includes getting/setting a specific sector
configuration, as well as getting/setting the selected
sector which is used to communicate with a station.

Change-Id: I9b10a9a580570df9fce63ddac171c1fce7bae4a1
Signed-off-by: Lior David <liord@codeaurora.org>