wil6210: prevent usage of incorrect TX hwtail
txdata->enabled is used in order to determine if the TX vring
is valid. As the data transmit is handled in a different context,
in case txdata->enabled is set before vring->hwtail is updated,
an old or corrupted vring->hwtail can be used.
Protect setting of txdata->enabled and vring->hwtail to prevent a
case where TX vring start handling TX packets before setting
vring->hwtail.

Change-Id: I32e98e24419b6d22ea06683e51ef0a320437c228
Signed-off-by: Maya Erez <merez@codeaurora.org>
Signed-off-by: Lior David <liord@codeaurora.org>