wil6210: extract firmware version from file header
Currently the FW version is taken from the sw_version field
of the FW ready event. This version is based on internal
version control revision and it is difficult to map to actual
FW version.
Fix this by using the actual FW version stored in the FW file
header record.

Change-Id: I58efea50ef87cab631201d593288d1c4586c6000
Signed-off-by: Lior David <liord@codeaurora.org>