mac80211: Set MAC address for FST only in the disable phase
The 11ad interface MAC address being updated twice in the network boot
sequence inside pre_mac80211() as part of FST configuration, once in
the "disable" phase and once in the "enable" phase.
As part of the update the interface goes down and after the "enable" phase
stay down.
Now the Mac address being updated only once, in the "disable" phase of the boot
sequence.

CRs-Fixed: 1079060

Change-Id: If799a6187ed277942840803fdada5f88a4131954
Signed-off-by: Alexei Avshalom Lazar <ailizaro@codeaurora.org>