wil6210: support configurable board file via sysfs
Add a sysfs allowing to override the board file containing
the RF characteristics of the 11ad device.
The new board file will be used the next time the interface
is brought up.
This change is intended to support cases where multiple
11ad devices exist in the platform. Each 11ad device can be
connected to a different antenna module and the HW cannot
automatically detect the type of antenna connected, so the
board file must be explicitly specified for each device.

Change-Id: I779735614b5752cc7f675fa86bdc44fa90dde61f
Signed-off-by: Lior David <liord@codeaurora.org>