mac80211: Fix possible sband related NULL pointer de-reference
This fixes possible potential NULL pointer de-reference possible in
a multivif scenario where 'sband' can be NULL, resulting in a crash
while acessing its members.  Fix this by replacing the existing API
'ieee80211_get_sdata_band' (which returns default 2GHZ  band even if
the channel context configuration is NULL) with  'ieee80211_get_sband'
which diligently returns a NULL pointer if the  channel configuration
is NULL. This change fixes the crash in single band cards (5Ghz only).
An example scenario is as below:

In multivif mode AP + STA in drivers like ath10k, when we
do a channel switch in the AP vif (which has a number of clients
connected) and a STA vif which is connected to some other AP,
when the channel switch in AP vif fails, while the STA vifs
tries to connect to the other AP, there is a window where
the channel context is NULL/invalid and this results in a
crash  while the clients connected to the AP vif tries to
reconnect and this race is very similar to the one investigated
by Michal https://patchwork.kernel.org/patch/3788161/ and this
does happens with hardware that supports 5Ghz alone over a very
long with continuous channel switch for the AP vif in regular
intervals, with wifi down/up

ieee80211 phy0: channel context reservation cannot be finalized because
some interfaces aren't switching
by Michal https://patchwork.kernel.org/patch/3788161/ and this
does happens with hardware that supports 5Ghz alone over a very
long with continuous channel switch for the AP vif in regular
intervals, with wifi down/up

ieee80211 phy0: channel context reservation cannot be finalized because
some interfaces aren't switching
wlan0: failed to finalize CSA, disconnecting
wlan0-1: deauthenticating from 8c:fd:f0:01:54:9c by local choice
      (Reason: 3=DEAUTH_LEAVING)

  WARNING: CPU: 1 PID: 19032 at net/mac80211/ieee80211_i.h:1013 sta_info_alloc+0x374/0x3fc [mac80211]
  [<bf77272c>] (sta_info_alloc [mac80211])
  [<bf78776c>] (ieee80211_add_station [mac80211]))
  [<bf73cc50>] (nl80211_new_station [cfg80211])

  Unable to handle kernel NULL pointer dereference at virtual
  address 00000014
  pgd = d5f4c000
  Internal error: Oops: 17 [#1] PREEMPT SMP ARM
  PC is at sta_info_alloc+0x380/0x3fc [mac80211]
  LR is at sta_info_alloc+0x37c/0x3fc [mac80211]
  [<bf772738>] (sta_info_alloc [mac80211])
  [<bf78776c>] (ieee80211_add_station [mac80211])
  [<bf73cc50>] (nl80211_new_station [cfg80211]))

Change-Id: I2bc35c57f4d4983388dfc476259fd4a3d8e946dc
Signed-off-by: Mohammed Shafi (Mohammed Shafi) Shajakhan <mohammed@codeaurora.org>