mac80211: fix fallback to macaddr
If '/sys/class/ieee80211/phyN' is not a symlink, the call to readlink
produces either an empty path (if /sys/class/ieee80211/phyN/device doesn't
exist) or a path that begins with '/sys/class/ieee80211' (which does not
work with the subsequent line to strip '/sys/devices' from the beginning of
the path).  Modify the checks to detect these conditions, and fallback to
using the macaddr.

Change-Id: I461b7f6b0c9f1714667ce503a44d6f3701c3a11e
git-svn-id: svn://svn.openwrt.org/openwrt/trunk@44172 3c298f89-4303-0410-b956-a3cf2f4a3e73
Signed-off-by: Nathan Hintz <nlhintz@hotmail.com>
Signed-off-by: Mohammed Shafi Shajakhan <mohammed@codeaurora.org>
Signed-off-by: Tamizh chelvam <tamizhr@codeaurora.org>