mac80211: Revert adding SM Power Save capability to hostapd configuration
Part of rev 44173 added setting the SM Power Save capability in the hostapd.conf
file if the driver indicated that it was supported. It appears this was
incorrect, because the field in the actual HT Capabilities field in the AP
configuration is really a state indication. Just copying the state from the
capability resulted in the AP indicating that it had SMPS enabled all the time
if it supported SMPS. This effectively just disables all clients from sending
packets to the AP with more than one spatial stream, for no good reason.
So remove this part of the change.

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@44239 3c298f89-4303-0410-b956-a3cf2f4a3e73

Change-Id: I772fddf8449ba89258a84c61e12a340eafca2efe
Signed-off-by: Robert Hancock <hancockrwd@gmail.com>
Signed-off-by: Mohamme Shafi Shajakhan <mohammed@codeaurora.org>
Signed-off-by: Tamizh chelvam <tamizhr@codeaurora.org>