ath10k: handling qos at STA side based on AP WMM enable/disable
Data packets are not sent by STA in case of STA joined to
non QOS AP (WMM disabled AP). This is happening because of STA
is sending data packets to firmware from host with qos enabled
along with non qos queue value(TID = 16)
Due to qos enabled, firmware is discarding the packet.

This patch fixes this issue by updating the qos based on station
WME capability field if WMM is disabled in AP.

Change-Id: I1299aa11d954a1633e4856baef5c6c8f45c5f033
Signed-off-by: Balaji Pothunoori <bpothuno@codeaurora.org>