mac80211: fix ath9k potential info leak
When the user sets count to zero the string buffer would remain
completely uninitialized which causes the kernel to parse its
own stack data, potentially leading to an info leak. In addition
to that, the string might be not terminated properly when the
user data does not contain a 0-terminator.

Change-Id: I4d811ed5628712e4e51d4d39b5ed1392168acb4e
Signed-off-by: Miaoqing Pan <miaoqing@codeaurora.org>