ath10k: fix spectral scan for QCA9984 and QCA9888 chipsets
In QCA9984 and QCA9888 chipsets, harware param value 'spectral_bin_discard'
is 12 bytes. This 12 bytes derived as segment index (4 bytes), extra bins
before the actual data (4 bytes) and extra bins after the actual data (4 bytes).
Right now, this 12 bytes discard from the end of the samples. This is causing
the wrong alignment of frequency samples. In every samples, we need to consider
the offset while taking actual spectral data. For QCA9984, QCA9888 the offset
is 8 bytes (segment index + extra bins before actual data). Introducing a
hardware parameter 'spectral_bin_offset' to resolve this issue in QCA9984 and
QCA9888 chipsets. For other chipsets, the hardware parameter
'spectral_bin_offset' is zero so that existing behaviour is retained as it is.

Change-Id: I83bbb799cfc159f35372ceaac2b3c0cdf5470f9c
Signed-off-by: Karthikeyan Periyasamy <periyasa@codeaurora.org>