mac80211: fix rx duration stats read after reset
Currently when per-station stats are reset, we reset
all the parameters(for pkts and bytes) but leave out
the duration. During stats read, we provide the
aggregated rx duration info. Hence to show the delta
duration after stats reset, offset duration is recorded
during a reset and this is used to get the delta when
the stats are read next time.

Change-Id: Ie3d1c5ecc8f1de977578a2efd371e47c520edfa3
Signed-off-by: Sriram R <srirrama@codeaurora.org>