mac80211: Assign correct dataif ops when setting 4addr mode separately
When 4addr mode is set sperately from userspace after an interface is brought up
as station, the 80211 encap offload flag/ops which is set during iface
bringup is not reset and this results in a WARN_ON during data
transmission using ieee80211_dataif_8023_ops. This is avoided by
setting the correct flag/ops when encap offloading is not supported during
an interface set/change.

Change-Id: I98f0419a9e2fef6910f867c07cb911d41f5487dd
Signed-off-by: Sriram R <srirrama@codeaurora.org>