Revert "mac80211: reject/clear user rate mask if not usable"
This reverts upstream commit e8e4f5280ddd0a7b43a795f90a0758e3c99df6a6

If the user rate mask results in no (basic) rates being usable,
clear it. Also, if we're already operating when it's set, reject
it instead.

Technically, selecting basic rates as the criterion is a bit too
restrictive, but calculating the usable rates over all stations
(e.g. in AP mode) is harder, and all stations must support the
basic rates. Similarly, in client mode, the basic rates will be
used anyway for control frames.

This fixes the "no supported rates (...) in rate_mask ..." warning
that occurs on TX when you've selected a rate mask that's not
compatible with the connection (e.g. an AP that enables only the
rates 36, 48, 54 and you've selected only 6, 9, 12.)

Change-Id: If88deedf1c829f82a3e3cb8372a6c5247b5eee03
Signed-off-by: Maharaja Kennadyrajan <mkenna@codeaurora.org>