wil6210: support VR mode
Allow wil6210 to configure the system in Virtual Reality (VR) mode.
When system is in VR mode it should have a low latency, and a high
responsive to events in the system. Wil6210 will set the rx interrupt
moderation to lower threshold and will drop packets when the transmit
queue is full.

New sysfs "vr_profile" can be used by userspace to enable and set
VR profile.
WMI_SET_VR_PROFILE_CMDID is used to configure the FW into different
VR profile configurations.

Change-Id: If9e0a5eb3323f1479d76b259c848310508e7009c
Signed-off-by: Ahmad Masri <amasri@codeaurora.org>