mac80211: fix low throughput in push pull mode
If a station is not eligible for transmission in ieee80211_txq_may_transmit(),
no packet will be delivered to FW. During the tests in push-pull mode with
many clients, after several seconds, not a single station is an eligible
candidate for transmission since global time is smaller than all the station's
virtual airtime. As a consequence, the Tx has been blocked and throughput is
quite low.

To avoid this situation to occur in push-pull mode, the new proposal is:

- Increase the airtime grace period a little more to reduce the unexpected sync

- If global virtual time is less than the virtual airtime of any station, sync it
  to the airtime of first station in the red-black tree

- Round the division result since the process of global virtual time involves the
  division calculation

Change-Id: I1387f0029269a0d2de517b1fca38bca3611987d3
Signed-off-by: Yibo Zhao <yiboz@codeaurora.org>