ath11k: add support for controlling tx power to a station
This patch will add the support to control the transmit
power for traffic to a station associated with the AP.

Underlying firmware will enforce that the maximum tx power
will be based on the regulatory requirements. If the user
given transmit power is greater than the allowed tx power
in the given channel, then the firmware will use the
maximum tx power in the same channel.

Max and Min tx power values will depends on number of
tx chain masks. The allowed tx power range values are
from 6 to 23.

When 0 is sent to the firmware as tx power, it will revert
to the default tx power for the station.

Change-Id: I4e3045ba3cd07414c043d75d1b811a7244de51da
Signed-off-by: Maharaja Kennadyrajan <mkenna@codeaurora.org>