mac80211: fix use after free issue during wifi restart test
Txq usually get removed from rbtree under the condition that it
is on the rbtree and in the remove list as well. Yet in wifi
restart stress test, there is chance that txq is not in the
remove list due to data buffered in it. Then if this freed txq
is used elsewhere, it will cause used after free issue.

Checking txq in remove list during removal is unnecessary since
in normal case, it is usually in remove list.

Change-Id: I91c946712eee747e63d190fc15eae10e1346a2a3
Signed-off-by: Yibo Zhao <yiboz@codeaurora.org>