ath10k: Fix channel info parsing for non tlv target
The tlv targets such as WCN3990 send more data in
the chan info event, which is not sent by the non tlv
targets. There is a minimum size check in the wmi event for
non-tlv targets and hence we cannot update the common
channel info structure.
If the common channel info structure is updated, the
size check for chan info event for non-tlv targets
will fail and return -EPROTO and we see the below
error messages
   ath10k_pci 0000:01:00.0: failed to parse chan info event: -71
Add tlv specific channel info structure and restore the
original size of the common channel info structure to
mitigate this issue.

CR-Fixed: 2489751

Change-Id: I4241b8d8a19b5f85229a19cdd0604212a732dd2d
Signed-off-by: Rakesh Pillai <pillair@codeaurora.org>
Signed-off-by: Lei Wang <leiwa@codeaurora.org>