ath11k: avoid WMM param truncation and burst time conversion logic
In conf_tx() mac operation callback, we are truncating the tx
params cw_min and cw_max due to lower data type cast and
WMI_VDEV_SET_WMM_PARAMS commmand expects the txoplimit param
in the units of 32 microseconds. so no need to convert the txop
unit from 32 microseconds to absolute microseconds. Also modified
the data type of cwmin and cwmax to avoid the trucation issue.

Change-Id: Ie6cd3a8602cc1a61d79bfebd755ebc53574b0c34
Signed-off-by: Karthikeyan Periyasamy <periyasa@codeaurora.org>