ath11k: correct vht guard interval mapping
Guard interval value which comes from VHT_SIG_A TLV has a mapping
where value 0 corresponds to LGI and values 1 and 3 corresponds
to SGI and SGI with NYSM ambiguity respectively both of which are
SGIs.

Value 3 which is actually SGI(0.4us) in VHT was mapped to a GI of
3.2us(only applicable in HE) resulting in incorrect rx GI stats.

This patch corrects the mapping.

Change-Id: I5fbaaccda61041ebcb8956489b742ea298152ad0
Signed-off-by: Manikanta Pubbisetty <mpubbise@codeaurora.org>