mac80211: Enable 11N and 11W support for ath11k
Currently ath9k and ath10k has dependency on 11N and 11W
so 11N and 11W are getting enabled along with ath9k and ath10k
modules.
This change adds 11N and 11W dependency for ath11k too so that
when ath11k alone is enabled, 11N and 11W will also get enabled.

Change-Id: I7b7e8ac0cc4f87f20257768f0b34f6177a99fea7
Signed-off-by: Ram Chandra Jangir <rjangir@codeaurora.org>