ath11k: Add support to search the peer with pdev_idx
peer_create() is not allowed for the same MAC address in the same
radio. To detect/find the peer in the same radio, we need to store
the pdev_idx in the created peer object. so that we can find the
peer with the requested pdev_idx is already present or not.

Change-Id: Ie0f2ea8f79ad74c83ddc88d2a61b7d24fc72d2fd
Signed-off-by: Karthikeyan Periyasamy <periyasa@codeaurora.org>