ath11k: add pktlog checksum to trace events to support pktlog
Pktlog data are different among the chipset & chipset versions.
As part of enhancing the user space script to decode the pktlog
trace events generated, it is desirable to know which chipset/
chipset version has provided the events and thereby decode the
pktlogs appropriately. Pktlog checksum helps to determine the
chipset variant which is given by the firmware in the struct
wmi_ready_event. pktlog checksums are computed during the
firmware build. So adding pktlog checksum to trace events.

Change-Id: I87de7093be4ad8ab7f52768661696d8af185761f
Signed-off-by: Maharaja Kennadyrajan <mkenna@codeaurora.org>