ath11k: Reset the order flag in frame header for nwifi decap
The HT control header of the wifi header will be removed by the
hw during native wifi decap. But the order flag in the decapped header 
is not reset. Also the ht control header is not used by mac80211 currently.
Hence reset the order flag in the frame control field of header.
This will avoid wrong header length calculation leading to traffic stall.

Change-Id: I17c37c102978cf7da8f7567f74d61f82e60044f2
Signed-off-by: Sriram R <srirrama@codeaurora.org>