mac80211/ath11k: send HE operation param to firmware.
HE operation IE in beacons are configured based on userspace params,
which firmware might not be aware of, thereby not enabling certain
features with respect what is being advertised in beacons. This patch
sends HE oper param from beacons to driver and driver sends it to firmware
using WMI interface for configuration.

Change-Id: Ica1281d616946ae2faaff5a233cb3ea21573bdc9
Signed-off-by: Pradeep Kumar Chitrapu <pradeepc@codeaurora.org>