ath11k: add tx hw 802.11 encapusaltion offloading support
This patch adds support for ethernet rxtx mode to the driver. The feature
is enabled via a new module parameter. If enabled to driver will enable
the feature on a per vif basis if all other requirements were met.

This patch is taken from this link https://patchwork.kernel.org/patch/11043205/
which is in linux mailing lists for ath11k driver.

Change-Id: I75a515f9177ea25baa4d424fa038d99438f66d9c
Signed-off-by: Venkateswara Naralasetty <vnaralas@codeaurora.org>
Signed-off-by: Manikanta Pubbisetty <mpubbise@codeaurora.org>