ath11k: rx path optimizations
During RX, accessing the reo dest ring descriptor directly is consuming
a lot of CPU cycles. Accessing the descriptor after copying it locally
has improved CPU usage by around 8~10% in the case of DBTC test case and
the improvement was close to around 13~15% in the case of lower frame
size test cases.

Change-Id: I9e11be8ac589eaba4f7cb01e19431477e9bcf0ac
Signed-off-by: Manikanta Pubbisetty <mpubbise@codeaurora.org>