ath11k: fix reo flush send
we are sending the reo flush command for the deleted peer
tid after the ageout period reaches 1 second. This handling
causes reo ring get full when more than 128 clients are
disconnected continuously. so added the count for flush list
and reo flush command is triggered after the list count reaches
the threshold value, it is configured as 64 (half of the reo ring).
This will avoid the situation where reo ring get full.

Change-Id: Ic338114194f37072819437d604311e8dc6840fab
Signed-off-by: Karthikeyan Periyasamy <periyasa@codeaurora.org>