ath11k: fix parsing PPDU_CTRL type in pktlog
PPDU_CTRL type is missing in current pktlog dumps.
PPDU_CTRL is sent on CE5 with len 2560 bytes, current driver ignores the payload len
greter than 2048.
PPDU_CTRL of 2560 bytes is sent in two fragments of len 2028 and 532 bytes, but firmware
reports pkt header has length as 2560 for both of the fragments.

Change-Id: I039df20bbe8560ff47b6d8d5fda7526aac07eb62
Signed-off-by: Anilkumar Kolli <akolli@codeaurora.org>