ath11k: disabling credit flow for wmi path
We can do this by telling the fw by two ways. One is a global
flag(HTC_MSG_SETUP_COMPLETE_EX_ID msg type flags) and
local flag(ATH11K_HTC_CONN_FLAGS_DISABLE_CREDIT_FLOW_CTRL = 1 << 3)
so we are setting both of them for safer side.
We need to set the second bit of the global flag to disable
credit flow completely in our driver.

I have added a hw_param member for credit flow control by which we can
enable or disable it depending on our choice. Also it will be compatible
with the chipsets which we are going to support in future with ath11k.
currently by default IPQ8074 and IPQ6018 both have credit flow disabled.

Change-Id: I93382f6af2e863f02e274b445ae7b82d521d4df3
Signed-off-by: Pravas Kumar Panda <kumarpan@codeaurora.org>