ath: Modify ath_key_delete() to not need full key entry
tkip_keymap can be used internally to avoid the reference to key->cipher
and with this, only the key index value itself is needed. This allows
ath_key_delete() call to be postponed to be handled after the upper
layer STA and key entry have already been removed. This is needed to
make ath9k key cache management safer.

Change-Id: Ib6270d7ec836cf7f589e75a33727f54f7b03a989
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Signed-off-by: Ronghua Zeng <ronghuaz@codeaurora.org>