ath11k: monitor ring stuck
More than 20000 ppdu id jumping can cause status ring and destination
ring processing not sync. The status ring is processed and the
destination ring is not processed. Since destination is not reaped for
so long, backpressure is occured at destination ring.

Added a workaround to update the ppdu id with the latest ppdu, this will
reap the destination ring and will avoid stuck. Increased monitor ring
sizes to avoid the early backpressure.

Signed-off-by: Karthikeyan Kathirvel <kathirve@codeaurora.org>
Change-Id: I0ef74ddb3e76ad358effeee8e85b7121a5d5c373