mac80211: Mon co-exist is crashing in NSS Offload
In case of monitor co-existence ath11k is changing the AP interface
encap from ethernet to native to capture the own bss packets, with
this config crash is observed in nss offload case when the encap is
native and decap is ethernet.

As nss offload pkts are on fast path not required to change the encap
to native and so not changing the encap type in case of nss offload.

Signed-off-by: Karthikeyan Kathirvel <kathirve@codeaurora.org>
Change-Id: I9d05c08477c4e498c6ab33dd0bccced5ccf9100a