mac80211: Fix crash caused by uninitialized MBSSID list
When /etc/config/wireless file is used, the first created interface of
type 'AP' is always set as the transmitting interface and a list is
initialized to store pointers to all non-transmitting interfaces
belonging to the same multiple BSSID set.

However, when hostapd is started manually, this code path does not get
executed. Creation of the first non-transmitting interface results in
a crash due to non-initialized list.

This patch fixes this crash by moving the initialization to function
ieee80211_if_add() for all types of interfaces.

Signed-off-by: Aloka Dixit <alokad@codeaurora.org>
Change-Id: I2e5589d0014b3e99b2d0698635f32fbe3c649224