ath11k: send proper txpower, maxregpower values to FW
Set proper values for txpower, maxregpower, maxpower, antennamax
in wmi scan chan list and vdev start command.

If txpower is not set, it will lead to cca stuck resulting in
latency issues for pine.

Maxregpower, maxpower, antennamax values has to be sent as it is.
Firmware will convert power values to 0.5dbm steps by multiplying
it with 2.

Signed-off-by: Lavanya Suresh <lavaks@codeaurora.org>
Change-Id: Ie49b91cb98a32eb2698099198616f16653161d6a