ath11k: Allow fwlogs by default after device registration
Currently diag events are allowed only when there is an active
pdev which happens when ATH11K is in ON state. But during
ftm mode we dont move the ath11k driver state to ON.
But fwlog events are received once it is enabled after
mac registration. Hence just check for valid pdev before
passing on the fwlog data.

Also use a modparam for enable/disable fwlog

Change-Id: Id9cde974fb3ccfde7870755759808c94cceeebc7
Signed-off-by: Sriram R <srirrama@codeaurora.org>