mac80211: RNR for EMA enabled AP
This change uses offsets provided by application to store separate
blocks for RNR (reduced neighbor report) each corresponding to one
multiple BSSID element block in the kernel. Both are added to create
EMA (Enhanced multi-BSSID advertisements) beacons.
Application is responsible of creating blocks such that if a
non-transmitted profile is included in multiple BSSID element then
it is excluded from the RNR of the same beacon.
RNR block for each EMA may also include entries common for all beacons
such as co-located APs.
Offsets are not used if EMA is not enabled.

Signed-off-by: Aloka Dixit <alokad@codeaurora.org>
Change-Id: I31e79053329835436bf774570ea9c0b22ca7b5f2