ath10k: Validate first subframe of amsdu before processing the list
Validate amsdu list by checking if the first frame has a valid
DA. In certain scenarios a proper msdu can be received as an amsdu when
the A-MSDU present bit gets flipped.
Since this bit is unauthenticated the HW crypto engine can pass the frame
to driver without any error indication. This could result in processing
unintended subframes collected in the amsdu list.

Comparing the non-aggregated msdu and an amsdu, the fields of the first
subframe DA matches with the llc/snap header fields of normal msdu.
In order to avoid processing such frames, add a validation to
filter such amsdu frames where the first subframe header DA matches
with the llc/snap header pattern.

Change-Id: Ibc39a17426354a2ada501bc631bbd9e85e73ece4
Signed-off-by: Sriram R <srirrama@codeaurora.org>