ath11k: add debugfs to disable dynamic bandwidth operation
By default dynamic bandwidth operation is enabled. Setting this debugfs
to 1 disables it.

This config option is temporary WORKAROUND till the FCC requirements are
made clear. FCC regulation currently expects an AP to stop ALL its
transmissions in the operating channel when there is an interference
above CCA-ED threshold in any one of the interfering sub-channels.
This is only possible if we disable dynamic BW.

Signed-off-by: Pradeep Kumar Chitrapu <pradeepc@codeaurora.org>
Change-Id: I6b78b9c0d69a6f6d5f42b12bd8d049d153d4b9ff