ath11k: Increment mgmt_tx count before tx send
Updated to increment 'num_pending_mgmt_tx' count before
invoking send, to avoid 0 value while decrementing in tx
completion handler.

Warning trace is seen in tx completion handler that count
is already 0 while decrementing. It is suspected that tx
completion might be received before count is incremented,
and so added this change. It's not yet proved that it
fixes the actual issue.

Signed-off-by: Lavanya Suresh <lavaks@codeaurora.org>
Change-Id: Iab78ffc461981dfb2eb94907d7b3c9e6441933d0