ath11k: Free skb on unsupported packet types received on ext paths
This change is based on code review and doesn't fix any known issues.
Also added a warn for a similar case in wifili ext vdev to warn on such
unsupported packets.

There is a memory leak due to kfree of skb, it is changed to
dev_kfree_skb_any to cleanup the entire skb properly.

Signed-off-by: Sriram R <srirrama@codeaurora.org>
Change-Id: I361a9ebd8048bc68556beb9841d2ffc7124a41b4